servers: Minor fixes/changes to the merge in #10a694712393ec4adfe1224c2135afeee221fe1e
