+++ java/servers/src/org/xtreemfs/osd/rwre/ReplicatedFileState.java
@@ -25,6 +25,7 @@ import org.xtreemfs.pbrpc.generatedinterfaces.OSDServiceClient;
 public class ReplicatedFileState extends RedundantFileState {
 
     private List<ObjectVersionMapping> objectsToFetch;
+    private int                        numObjectsPending;
 
     public ReplicatedFileState(String fileId, XLocations locations, ServiceUUID localUUID,
                                OSDServiceClient client) throws UnknownUUIDException, IOException {
@@ -44,4 +45,20 @@ public class ReplicatedFileState extends RedundantFileState {
     public void setObjectsToFetch(List<ObjectVersionMapping> objectsToFetch) {
         this.objectsToFetch = objectsToFetch;
     }
+
+    /**
+     * @return the numObjectsPending
+     */
+    public int getNumObjectsPending() {
+        return numObjectsPending;
+    }
+
+    public void incrementNumObjectsPending() {
+        ++numObjectsPending;
+    }
+
+    public void decrementNumObjectsPending() {
+        --numObjectsPending;
+    }
+
 }
